connect to development database
