+++ src/db.dev.js
@@ -1,0 +1,10 @@
+
+// require('dotenv').config();
+
+module.exports = {
+    url: process.env.DEVELOPMENT_DB,
+    options: {
+      useNewUrlParser: true,
+      useUnifiedTopology: true,
+    },
+  }
